Milwaukee Thursday 7:37 -> End of the Night
At the current moment, we can reliably complete the center gear peg and
the baseline. The error value for DriveToPosition needs adjusted if we
are to complete the side pegs, but that is not to be adjusted unless
competition team wants the feature. Teleop works and the camera is
showing inconsistently. We can climb. This code is not to be changed
unless absolutely necessary. Everything works, so we are locking it in.